+++ code/jpf-ctl/src/main/java/ctl/AtomicProposition.java
@@ -20,7 +20,7 @@ package ctl;
 import java.util.HashSet;
 import java.util.Set;
 
-import config.Type;
+import label.Type;
 
 /**
  * This class represents the CTL state formula consisting of an atomic proposition.
