+++ code/jpf-ctl/src/main/java/label/Type.java
@@ -1,4 +1,4 @@
-package config;
+package label;
 
 public enum Type {
 	Initial,
